Tweak the base search view to not require any search help text
