+++ timetable_solutions/interfaces/data_management/forms/base_forms.py
@@ -1,39 +1,8 @@
 """Base forms that are subclassed or just used directly for managing individual models."""
 
-# Standard library imports
-from typing import Any
-
 # Django imports
 from django import forms
 from django.core.files import uploadedfile
-
-
-class Search(forms.Form):
-    """Form providing a single search field."""
-
-    search_term = forms.CharField(
-        required=True,
-        label="Search term",
-        initial="",
-        error_messages={"required": "Please enter a search term!"},
-    )
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        if kwargs.get("search_help_text"):
-            self.base_fields["search_term"].help_text = kwargs.pop("search_help_text")
-        super().__init__(*args, **kwargs)
-
-    def clean_search_term(self) -> str:
-        """Prevent single letter searches."""
-        if search_term := self.cleaned_data.get("search_term"):
-            try:
-                int(search_term)
-            except ValueError:
-                if len(search_term) < 2:
-                    raise forms.ValidationError(
-                        "Non-numeric search terms must be more than one character!"
-                    )
-        return search_term
 
 
 class CreateUpdate(forms.Form):
